fix(specs): restore `as u8` cast on MAX_SIGNERS for spl-token compatibility
spl-token's MAX_SIGNERS is `usize` while multisig.m/n are `u8`.
The cast is necessary for spl-token builds; p-token's MAX_SIGNERS
is already `u8` so the cast is harmless there.

+++ specs/shared/test_process_revoke_multisig.rs
@@ -9,10 +9,10 @@ fn test_process_revoke_multisig(accounts: &[AccountInfo; 3]) -> ProgramResult {
     #[cfg(feature = "assumptions")]
     {
         let multisig = get_multisig(&accounts[1]);
-        if multisig.m < 1 || multisig.m > MAX_SIGNERS {
+        if multisig.m < 1 || multisig.m > MAX_SIGNERS as u8 {
             return Ok(());
         }
-        if multisig.n < 1 || multisig.n > MAX_SIGNERS {
+        if multisig.n < 1 || multisig.n > MAX_SIGNERS as u8 {
             return Ok(());
         }
     }
